Fixes? I think

+++ src/main/Camera.java
@@ -3,7 +3,7 @@ package main;
 public class Camera {
     Vector location;
 
-    public Camera(double x, int y, int z) {
+    public Camera(double x, double y, double z) {
         this.location = new Vector(x, y, z, (byte) 1);
     }
